# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cyber Code API",
    "version": "1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4747"
    }
  ],
  "security": [
    {
      "basic": []
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "HistoryPage": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventEnvelope"
            }
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "data",
          "hasMore"
        ]
      },
      "EventEnvelope": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "data": true,
          "location": {
            "type": [
              "string",
              "null"
            ],
            "description": "The Location directory the event belongs to."
          },
          "durable": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DurableRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "type",
          "data"
        ],
        "description": "One event on any stream."
      },
      "DurableRef": {
        "type": "object",
        "properties": {
          "aggregateID": {
            "type": "string"
          },
          "seq": {
            "type": "integer",
            "format": "int64"
          },
          "version": {
            "type": "integer",
            "format": "uint32",
            "minimum": 0
          }
        },
        "required": [
          "aggregateID",
          "seq",
          "version"
        ],
        "description": "Where a durable event sits in its aggregate's history."
      },
...
      "ErrorBody": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "title": "ErrorBody",
        "description": "The JSON body of every declared failure.",
        "type": "object",
        "properties": {
          "_tag": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "ref": {
            "description": "`err_<id>` for `UnknownError`, logged server-side.",
            "type": [
              "string",
              "null"
            ]
          },
          "paths": {
            "description": "Paths for rewind conflicts.",
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "_tag",
          "message"
        ]
      }
    }
  },
  "paths": {
...
    "/api/v1/event": {
      "get": {
        "operationId": "v1.event.subscribe",
        "summary": "Live events for a Location (scope=all for every Location)",
        "tags": [
          "event"
        ],
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "location[directory]",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-cyber-directory",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events; each `data` is an EventEnvelope",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sessions/{sessionID}/history": {
      "get": {
        "operationId": "v1.session.history",
        "summary": "Durable events after a sequence",
        "tags": [
          "session"
        ],
        "parameters": [
          {
            "name": "sessionID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPage"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessions/{sessionID}/events": {
      "get": {
        "operationId": "v1.session.events",
        "summary": "Replay durable events, then follow",
        "tags": [
          "session"
        ],
        "parameters": [
          {
            "name": "sessionID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server-sent events; each `data` is an EventEnvelope",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/EventEnvelope"
                }
              }
            }
          },
          "default": {
            "description": "Tagged error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  }
}